+++ crates/krb5-client/src/bin/krb5-kvno.rs
@@ -167,6 +167,7 @@ fn run(
         cname: cred.client.1.clone(),
         crealm: cred.client.0.clone(),
         fast_avail: false,
+        pa_type: None,
     };
     let tgs = if let Some(who) = for_user {
         let (uname, urealm) = if who.contains('@') {